Move editor A/B test bucketing code from PHP to JS, don't use cookies
Previously the token to follow anon users across sessions was stored
in a cookie, since that was the only way we could do it in PHP.

Turns out that setting a cookie disables caching of the response. Oops.

Now it's stored in localStorage. We have to manually clear it after
90 days, since localStorage entries never expire by themselves.

The token is now generated using mw.user.generateRandomSessionId().
It's shorter than before, but this function is meant for exactly this
purpose, so I assume the length is sufficient to avoid duplicates.

Bug: T236337
Change-Id: I1b5af75fe6da720b5a15adadd1a27d7baca178ff