Allow block message drawer to handle different block creators
Before this, the block message drawer assumes every block is
created by a local user and tries to link to the user's page.

However, some blocks are created by users who are not local to
the wiki, and some are created by the software (e.g. if an IP
address is blacklisted in a wiki's config).

The API provides the block creator's name and their local user
ID. After this commit, the block drawer decides what to display
about the block creator based on this information:
* If the creator has a nonzero ID, they are a local user, so
  link to their page
* If the creator has a name but has an ID of 0, they are not a
  local user, so display the name but do not try to link to a
  user page
* If the creator has no name, the block was created by the
  software, so do not display creator information.

Bug: T236814
Change-Id: Ie2aa00cfec5e166460bcaeddba3c601fe0627c13