MobileContext: Remove unused method `getForceMobileView()`
It was intended to get the $this->forceMobileView value but was never
used. The only usage of $this->forceMobileView that meets this use-case
is on L#394 (in the same class). Can it be safely removed as the method
is not used anywhere else. See usage below;

https://codesearch.wmflabs.org/search/?q=getForceMobileView&i=nope&files=&repos=

Or should we replace the $this->forceMobileView check on L#394 to the
method call instead?

Change-Id: I17bbf5c6de9f761b5d976812981cd5d01686ece3