Correct the back button behaviour for talk overlay with unstaged changes
When entering a reply in a talk section overlay and hitting back, the URL
in the address bar is lost regardless of whether you click 'ok' or 'cancel'
the address bar should only change when the edit is abandoned.

Replication steps:
* Login
* Visit a talk section directly e.g. Foo#/talk/1 or Talk:Foo#Heading
* type a reply in the talk box
* click browser back button
* I click cancel

Expected: I see a confirmation prompt about abandoning edit and when
I exit the hash fragment is not '#'

Previously - the hash fragment was reset to '#' as soon as you click back.

Bug: T230695
Change-Id: I6521697d775644913325608e45dbbf6b10ae347f