Cleanup OverlayManager _matchRoute
This cleans up the work from
I80201394fd7015db6700446142b0b4b20829f12e which added support for string
literal routes passed to OverlayManager. We don't want the
`path.match` line to run if passed a string literal as it could be an
invalid regex and cause a JS error to occur (run modified test before
and after this change for proof).

Therefore, this patch makes it only execute if the
route is not a string (assumed to a regex).

Additonal changes:

* Modified test to catch this in the future

* Used `didMatch` in ternary for improved readability

Bug: T230695
Change-Id: I43d429b7a4366e24958aa3e61c155b45f0a7006a