Avoid unintentional back behavior in _onHideOverlay
Previously when hiding certain overlays, browser back was utilised
to ensure the address bar was correct.

However, this doesn't work when you hide the overlay by different mechanisms
e.g. setting the hash to '#'

A slight adjustment is made to the local to check the currently open hash
fragment.

Additional changes:
* Lengthy comments describe behaviour
* A function is renamed to help understand it better
* OverlayManager test coverage = 100%

Bug: T237677
Change-Id: Ieec8019facd19c9d9b97adb3f2435e0ef292962a