CURL as part of build step for docs
Having all these files inside MobileFrontend is confusing as they
do not live in production. This means they are greppable inside
codesearch.wmflabs.org and confusing to developers when changing them
has no effect on the end production.

Previously this happened because CURL was restricted in the doc scripts.
This should now be relaxed.

Bug: T213223
Change-Id: I3ca3800b9d74d2ca5513647d39df0400bb33a85b