Drawer is passed onShow function
This helps centralise the logic for displaying the Minerva
page mask to Minerva which aids the new main menu. It also
allows us to remove drawers from the DOM on close.

This makes the Drawer a little dumber while keeping
backwards compatibility.

Changes:
* Access to the window is limited to consumers which do not
pass the onShow event and when that happens a warning occurs
* Classes on parent are no longer a concern of the Drawer
* The closeOnScroll behaviour is moved upwards (in a more
Redux friendly direction)

Bug: T214045
Change-Id: I960872c4f827c2b5a72f0c0d0ee0946792fb6e3b