Drawer's no longer touch the body tag and have their own mask
Drawer is just a dumb component with a show and hide method.
The drawer component now contains the existing drawer as a child as well
as a mask omponent which covers the screen and when clicked hides the
drawer. This means that they are not tied to the Minerva mask that is
used for other purposes and that the component itself doesn't need to
reach out and touch the body tag.

All known drawers have been updated to be compatible with this change -
red link drawer, block drawer and ctas.

Bug: T214049
Bug: T214647
Change-Id: Ib123efe6af1b167706a73c71c860c85e7f439cc5