Remove mobile.site and mobile.styles modules
Now we have hooks in core, we can alter the pages that the site
modules use, rather than replacing them with entirely different modules.

These modules are retained for cached HTML with deprecation notices but
serve as empty.

Given we are now using the site module, $wgAllowSiteCSSOnRestrictedPages and
possibly other configuration variables begin to work with no additional cost.

Bug: T140045
Bug: T237050
Depends-On: I5166797d9b02346eda09ea86ad4056e466a891d3
Change-Id: I26644f49dff57249ebd7eb20818add8a7110de6b