Dev: rlModule BE GONE! OverlayManager routes must always be synchronous
Previously we allowed asynchronous overlays. This was problematic as it
didn't allow us to render leaving visual displays in rendering.

Now no overlays use it so we should not allow this. OverlayManager
routes must always return an overlay. For migration path, a
loadingOverlay can be used just as the editor overlay does things.

Bug: T214641
Depends-On: I73cb622bbda44f4cfe51d08189419e15003b9d91
Change-Id: Id0922af6dbe5726d6b93b11e797295afa783d8d3