Remove redundant variable name from @var documentation
There are two usages for the @var tag: For class properties, and inline.
The inline form requires to repeat the variable name. But when documenting
class properties, this is not only redundant, it can potentially create
strange documentation where the property name shows up two times.

Other changes done to this code the same time:
* Prefer the "bool" and "int" shortcuts.
* Use "self" to reference the own class name, instead of repeating it.
* Remove a few comments that do not add any information, e.g. the fact
  that a $revision variable holds "the revision object".

Change-Id: Ieed077d153a76da9d84ca6fb2162a728744ded3a