Fix ReferencesHtmlScraperGateway showing child references
This is split from the patch I7f36653. This patch here is mostly
independent from the ongoing BookReferencing project. Instead of passing
the entire <li> element around (which contains elements that are then
hidden via CSS), we find the relevant "reference-text" element as early
as possible, and pass only this around. Note this CSS class name is
hard-coded in the Cite extension and guaranteed to be there.

This also fixes an actual bug that currently shows up on the beta cluster.
Go to
http://de.wikipedia.beta.wmflabs.org/wiki/Faultier?mobileaction=toggle_view_mobile
and click the [1]. Only the parent reference should be shown, but instead
all child elements are also shown. This patch here fixes this.

Bug: T242899
Change-Id: I677beddb461046b7f663a055f57f8690ff292a9c