mobile.init: Remove no-op 'timeStamp' compat
This seems to have been copied from ve.init.mw.trackSubscriber.js
in the VisualEditor repo. However, the code there was using
ve.track(), whereas this is using mw.track().

In 2015, ve.track() was given a third 'timeStamp' parameter, which
the VE code was providing compat for. However, this isn't relevant
in the MF code here.

Change-Id: I496042af53db26d3d6947ac940051557527da9c5