mobile.init: Migrate edit steps from this.timeStamp to mw.now()
Last user of this property. Once fixed, the logic and code for
precomputing this on all events can be removed, and after that
the code for mw.track()/mw.now() itself can be moved down from
'startup' module to 'mediawiki.base' for faster page load times
and fewer bytes on repeat downloads.

Bug: T229577
Change-Id: Ic4df07ef8ae0c1818089e12ee5339436fe9ad1a0