Add .mw-parser-output to citation drawer so that TemplateStyles apply
All styles written by editors using the TemplateStyles extension are
limited to only applying inside `.mw-parser-output`, to prevent them
from affecting the page layout.

When article content is pulled from its normal place and displayed in
popups, drawers, and so on, the class has to be added so that the
styles still apply. We do this in many places in VisualEditor as well.

Bug: T244510
Change-Id: I7ef1f7802e28c93239bca5fdcb25943f885391e6