Remove 'skipPreview' option from MF editor
This is never used in production, and skipping edit summaries
should not be encouraged. As such this is just unnecessary
complexity to maintain.

As this is the only property of MFEditorOptions, remove that
config too.

Note that the params passed to EditorOverlays are sometimes
called "editorOptions" locally, but that is not related
to MFEditorOptions.

Change-Id: I790927b68807418a307d0e886766fdcc4615121e