mobile.init: Remove unused code in apply2()
This private function is only used to combine two jQuery on()
event handlers.

Returning an array from an event handler is unexpected,
and does not do anything as far as I know (the only meaningful
thing an event handler can return is boolean false, which is a
legacy shortcut for e.preventDefault().

Change-Id: Idbacfdd4ec1feb86afc79d3294c8c87b4941e86a