Migrate 'user.tokens' dependency to 'user.options'
The 'user.tokens' module is a deprecated alias to 'user.options'
since MW 1.35.

Also, it seems the dependency was declared on the wrong module.
It used to be on "mobile.init", which does not contain any
use of mw.user.tokens. It looks like the only file in this repo
that uses it, is mobile.startup/amcOutreach.js, so I've moved
the dependency to that module instead.

Bug: T235457
Change-Id: I345d9d8d347b9c4d1b9cd9d46a7935a2a52ed2d8