[dev][fix] use NPM pathing when running bundlesize
check_bundle.sh assumes it will be invoked by package.json with an NPM
PATH environment. This means you can't run it directly without the
bundlesize executable in your PATH.

The issue is resolved by wrapping the `bundlesize` invocation in npx.
Since [npx will auto-install by default], disable that behavior for
safety since the assumption is that the script will be invoked from the
MobileFrontend repo.

Unfortunately, CI does not have npx so the following hack is used
instead:

  PATH="$(npm bin):$(npm bin -g):$PATH" bundlesize

That is, get the local package.json executable directory, get the global
bin directory too, and prepend it to PATH.

[npx will auto-install by default]: https://github.com/npm/npx/issues/9

Change-Id: I2e13633b1394ad7d7f2ede56416ca932a82a63db