specials: switch public methods to protected
Follow-up on Ib241a65ef9024a7d2753ef3a991f269f6a7c4de5

Make method protected as the original intension was to allow this to
be used by sub-classes only and not allow the API to be public for
public consumption.

Introducing it as public in the first place was a mistake. Sorry!

Change-Id: I8d99bb4ead865f6504d787202355d143ded9bfe3