LanguageInfo Overlay: Preprocess data for better UX
The previous version of this allowed for rendering languages even if no
autonym was present. This can theoretically work for the LanguageOverlay
which shows language links for titles; but since titles are not rendered
in the LanguageInfoOverlay, the layout ends up looking broken if
autonyms are missing as well.

This change also filters out legacy language codes.

Finally a comment in LanguageInfoOverlay has been removed, since it was
partially incorrect: the CLDR extension provides the autonyms used by
the languageinfo API, and when this is present these autonyms can be
used for search / filtering in the language info overlay.

Follows-up: Ic0f1613e8fd27788f67ed6235386491f0fe077c3
Change-Id: I5ae77a89c8d448641082002e09452011bf2516dd