Fix wrong section name in edit summary when editing whole page in source mode
When editing whole page, the edit summary erroneously included the
name of the first section on the page.

The mobile wikitext editor will usually only have the contents of a
single section available for editing. However, when switching from
visual mode with changes, we open it in a special mode where you're
editing the whole page (because editing e.g. references in visual mode
can affect wikitext outside of the section). There's also an open task
to allow editing the whole page in other scenarios (T203151).

Bug: T258832
Change-Id: Ib00d47238683a96a32b60b30c23ef694c3e86be1