ApiMobileView: Remove debug logging from parseSectionsData
ApiMobileView::parseSectionsData relies on an assumption that a split
on section heading elements yields the same number of sections as the
parser thinks there are in the page. On pages that break that
assumption, a debug message is logged, and section processing bails
out. ApiMobileView is deprecated, and no one is looking at these
messages. This patch removes the logging.

Bug: T260351
Change-Id: Idcb937cbb09276f55aca2fda821d799345fcb30d