LanguageInfoOverlay: Add route that excludes suggestions
Modify the languageInfoOverlay to only show "all languages" instead of
separating into suggested / other languages. For the context of a
generic language selector, sometimes hiding the suggestions makes
sense (as in T232410); other extension code might find it useful to
retain the suggested / other split, which is why this patch adds a
separate route so that both are supported.

Bug: T232410
Change-Id: Icce2631b2d5d80defc2287fb83f58f97a82a1b1e