Flatten Section transforms
In order to simplify the approach to enabling/disabling transform features,
it's good to have them all as array and set it up with required features only.

This patch moves LazyImageTransform and MoveLeadParagraphTransform out
of MakeSectionsTransform and make them equivalent to other transforms in the system.

NOTE: Both LazyImageTransform and MoveLeadParagraphTransform require
<section> tag to be added by MakeSectionsTransform and therefore must
follow it in the list

Bug: T261769
Change-Id: Iac1812ea0804e91150939abfe1dda725b52422f2