Simplify lead paragraph check
We are using two different approches for checking the emptiness
of a lead paragraph excluding TemplateStyles or coordinates.

As we are cloning the node now, the easier approach is just
to remove the tags we don't care about and check for emptiness.

This fixes the case paragraphs containing TemplateStyles were
never considered lead paragraphs.

Change-Id: Ifa7e6be8f070d337ad5234d11c253d75810ec40d