Remove usage of deprecated ParserCache::getKey.
ParserCache key should not be used for persistence.
The ParserCache::getKey method is being deprecated,
and the actual value of the key from now on is considered
internal to the ParserCache implementation.

This change will invalidate MobileView cache, however
given that the cache has a 1-hour TTL, I believe that
should not be an issue. Happy to make the change more
complex and transition the existing cache if you think
loosing it would be a problem.

Depends-On: I4a71e5a7eb1c25cd53b857c115883cd00160736b
Bug: T263689
Change-Id: I20b5a3eece79afaac6a4fef733d7a60ea23c6ffe