Don't show the mobile option when sections are already expanded
The setting on special mobile options expands all sections on pages
via JavaScript. If the config flag $wgMFCollapseSectionsByDefault
is disabled then the sections will already be expanded by default and
not only will this setting not make sense, it will break.

Historically this was disabled
on wikidata.org and caused some confusion, so for completeness we should
not show this.

Bug: T239195
Change-Id: I56864b4ac5888e33cf9a83b3026e93a108d26f8f