Emit editor-loaded event when editor has actually loaded
The editor-loaded event was emitted before loading was actually
completely finished. Right after emitting the event, we called
scrollToLeadParagraph(), which changes the selection. This made the
editor-loaded event useless for callers that need to change the
selection once the editor has loaded, because their selection change
would immediately be undone.

Instead, emit the editor-loaded event after scrollToLeadParagraph() and
all the other odds and ends. Codesearch found no users of this event, so
this is not a breaking change.

Change-Id: I3490f432aa04c8afa526fbdf75e56e8d2c3aee94