Fix signature of MockWikiPage::getLatest
Core change Ib3c9b22e0c0b7c97e is going to add an optional parameter to
WikiPage::getLatest(). This will break MockWikiPage::getLatest() if it
does not accept that parameter. However, PHP doesn't complain if the
subclass already has this parameter before the base class does. So this
adds it to MockWikiPage, for forward-compatibility with the core patch.

Needed-By: Ib3c9b22e0c0b7c97e6c230f2fecf0663b4c68cc6
Change-Id: Ia34cc8fa795b6b27832982fc22fa237ee36339b4