postEditMobile hook: Pass newRevId
This is useful if subscribers to the hook want to know if a new revision
was made during the save, or if a null edit occurred.

note, this is now bringing postEditMobile a bit farther away from how
the postEdit hook in core works (that one passes messages instead of
data), but it seems more useful to me to pass relevant data.

Bug: T269659
Change-Id: I17f17b45df013c0bfbd460b9584446d3d4b367cb