specials: Make SpecialContributions be able to view IP editors contribs
In this patch, an issue is resolved when non AMC mode is used where by it's
not possible to view contributions of IP editors. Inject the required services
via the contributions special page and pass it to parent (history).

NOTE - When the user is an IP, null or false is returned making the special
contributions page to error out instead of handling the error and rendering
a page with no contributions.

Bug: T291233
Change-Id: I8de22997772e09d38900de1ca2d374a395285293