Remove unwanted space from language search results
When user searches for a language inside language searcher, and
suggested languages exist, an extra unwanted margin appears at
the top of the search results container.

This is caused because of the existence of marker pseudoelements,
inside the (otherwise hidden) list items of the suggested languages
list. This only happens for suggested languages list (and not for
the all-languages list) because this is an ordered list (ol) and
list-style property has not been set.

This patch sets the list-style CSS property for order lists (ol)
that have the .site-link-list class to "none", to remove this
unwanted space.

Bug: T300592
Change-Id: Iafe567a4953a8abcf6be552f0eb0ab727ab45561