ParserOutput::getPageProperty() now returns null when key is missing.
The return value of ParserOutput::getPageProperty() has transitioned
to returning `null` instead of `false` when the page property is missing.

Took the opportunity to tweak MobileFrontendHooks::findTagline() so
that it also returns `null` if the tagline isn't found.  (The code
was written to expect falsy, the test cases expect false, and the
phpdoc said the return type was always string.)

Bug: T301915
Depends-On: Iaa25c390118d2db2b6578cdd558f2defd5351d15
Change-Id: I1128842f9cd9c14e50d1239c021e509d8537d238