LanguageSearcher: Fire a hook when LanguageSearcher view is rendered
This patch adds the required API inside LanguageSearcher
template to display the missing languages panel for the new Section
Translation entrypoint.

It also adds a "addBanner" method inside LanguageSearcher, which
accepts the message to display inside the banner.

Finally, the patch adds the "mobileFrontend.languageSearcher.onOpen"
and "mobileFrontend.languageSearcher.onBannerClick" hooks that is
being fired when the LanguageSearcher view is rendered, and banner
clicked respectively inside languageOverlay.js.

The onOpen hook is fired with the language searcher component as
argument.

Bug: T298032
Change-Id: I335547be1a0ac83bd1d5137b6d4d7dc067a4425e