Fix search result ordering
Fix implementation of the comparison function for sorting of
search results received from the API: The current implementation
is very fragile: anything which does not compare to be `<` is
given the +1 result (meaning “reverse the original order”). And
because the search results on Wikidata do not include an `index`
field at all, the sorting functions uses `undefined` as all keys.
And since `undefined < undefined` is false, all search results
on Wikidata end up being reversed by this “sort”.

Changed the implementation of the comparison function so that
it works correctly even for undefineds (also: the current
implementation does not even make a consistent comparator,
since it returns +1 for equal items instead of 0, meaning the
final sort order is implementation-defined, anyway): since
`undefined-undefined` is `NaN`, which is understood to be 0
by the sort comparator, so that in the Wikidata case, the
sort does not change the ordering of the items, displaying
the (correct) original ordering.

Bug: T307601
Change-Id: I16ea9079c0eed951197821b76a75e0c7b58ff2f4