Toggler: Fix revealing sections with special characters like `"` `,` `(`
The existing code only handled `.` and `:`, but there are other
characters that can't be used in selectors. Use the standard function
for escaping them.

Only the ID is now passed to reveal(), without the leading '#',
otherwise it would get escaped. One of the calls to reveal() already
did that, which didn't work previously, but it wasn't needed anyway.

Bug: T301932
Bug: T230293
Bug: T300620
Change-Id: I4ac18d0ad92a5867237b02360b4298a06fdd782d