Make sections expandable by find-in-page
This patch replaces the stylesheet to make toggleable sections visible
or not with the hidden=until-found attribute and beforematch event,
which makes toggleable sections revelable by the browser's find-in-page
and scroll to text fragment features.

Since the hidden attribute still applies display:none in browsers which
don't support hidden=until-found, no special casing is needed for any
browsers.

Bug: T239295
Bug: T216789
Change-Id: I9da1526f02ca9d3adc93b87173c6e7d8047136cc