Disallow using the `self` global variable
This extension commonly uses the pattern `var self = this` to access
the "outer" `this` inside of callbacks. Forgetting that line causes
things to break horribly, but it is not reported as a lint error,
because `self` is also the name of a global variable:
https://developer.mozilla.org/en-US/docs/Web/API/Window/self

As we have no code that uses it in this extension, disallow referring
to it, so that mistakes like in I83ae60d5d7 are flagged as errors.

Bug: T314737
Change-Id: I5949fcd4701ab1a68374a79a41490ba37cab1a6a