Compatibility with new heading HTML (editor)
Small fix to support new HTML markup for headings, which will soon be
used by Parsoid page views (T269630) and by the old parser (T13555).

When starting an edit from a section edit link, the viewport was not
scrolled to align with the clicked heading and thus didn't provide a
smooth transition into edit mode.

This is similar to change I89f8abac52 in VisualEditor.

Note that the Parsoid markup has other problems: T269499, which are
not addressed here.

Bug: T13555
Change-Id: If71d4d8292daa7e13a94521da8f9d9e5431f2f01