Reinitialize edit links after page content is reloaded
Previously, if the page content was reloaded without refreshing
the page (e.g. after a visual editor or a reply tool edit), the
event handlers that open the editor were not registered again on
the section edit links.

This was not a noticeable issue in most cases, because the links
have a 'href' attribute for the no-JS fallback, and we loaded the
editor in JS after following the link.

However, it prevented us from simulating clicks on the links in
DiscussionTools, where we offer a fallback to the normal wikitext
editor.

Bug: T324686
Change-Id: Ied8415f8dfa17a9fe3647859e88435748125ab71