Don't show a flicker of WikiEditor before displaying the MF editor
Same approach as VE employs; display a placeholder message instead as
the editor loads. Minor benefit is that we save on some data transfer
because the full page's wikitext won't be loaded only to be discarded.
Message includes a fallback so that mobile users with JS disabled or
with insufficient support can redirect to the old behavior.

This only happens when directly visiting an editor URL with action=edit.

Bug: T334263
Change-Id: I6028fa4a34916a6cd20161009cf5893492286c2d