Clean up mobile editor initialization checks
After d53979c050ade382e05263ea0508cd86ad84006a the JS and PHP code
sometimes did not agree whether the mobile wikitext editor should be
available on a page, leading to bugs like T334799.

Instead of trying to keep them consistent, move all of the checks to
the PHP side, and don't even load the editor code if it's not supposed
to be available.

Bug: T334263
Change-Id: I480c44417db85edf29e4af0a1ebcd072e913539a