editor: Use core post-edit notifications
The core post-edit notification module is available on mobile since
MediaWiki commit Ie0a49e5f37ac5e47066ce74d7e618d933cb30809, and was
already used for mobile edits not involving the mobile editor (e.g.
when undoing an edit). It is used consistently now.

This will also allow the mobile site to display notifications about
the creation of a temporary account (part of the IP Masking project),
which are implemented in the post-edit notification code.

Depends-On: I7b76b0f97bc4b5ca85f4b47bd27ae7568bcbe353
Bug: T336001
Change-Id: I1aff8937b195e5253ff4e3b68b209c7d33bfee10