editor: Load default content for 'MediaWiki:' namespace pages
When editing a 'MediaWiki:' namespace page that doesn't exist, the
default content of the localisation message should be preloaded.

EditorGateway no longer needs to know whether we're creating a new
page, it can figure it out from the API response.

Also improve the interactions with preloaded content in general:
allow saving the page immediately when non-default content is present.

Change-Id: I66cda3c5f5b890c4fd8bbd3d0cfe58daf46eb723