Opt out of stable policy for MobileFrontend hooks, remove one, limit mw usage
https://www.mediawiki.org/w/index.php?title=Stable_interface_policy%2FFrontend
makes hooks stable. The MobileFrontend hooks are meant to be internal
so these have all been documented as such.

While reviewing I also removed the
mobileFrontend.languageSearcher.onBannerClick hook as it was not
being used by any code or gadget so per the stable policy can be
removed without deprecation.

Bug: T348081
Change-Id: I8725f7dc7cb3e3ecff915dd3b901b6981d59a5f1