Handle mobile URLs for other wikis
MobileContext::getMobileUrl() breaks when used with an URL for
another wiki in the wiki farm, which is often necessary for
cross-wiki links. Fix it by replacing the mobile template (which
has a per-wiki value, so we'll always end up applying the template
for the current wiki to the URL even if it points to a different
wiki) with an arbitrary callback that takes the desktop domain
and returns the mobile domain.

Deprecate the old system ($wgMobileUrlTemplate). This also
deprecates using mobile URLs which differ in parts other
than the domain; this was never used and doesn't quite work
anyway (getDesktopUrl() ignores the path part).

(Note that the patch results in some changes in behavior
for URL templates with a path component. Since these were never
properly supported, it doesn't seem worth the effort to avoid.)

Also deprecate the GetMobileUrl hook which is undocumented, not
used anywhere, tied to the mobile template system, and didn't
seem particularly helpful in the first place.

Bug: T257852
Change-Id: If9100f2a9b86bd145c37745de511b1632f0acaa1