Replace usage of UserLoginComplete with PostLoginRedirect hook
With the changes in 5d2c1b05b (I8a2c91e1f173cf7b36f) that fixes
not calling the UserLoginComplete hook with $direct=false, this
patch makes MobileFrontend to use the appropriate hook for adding
articles to a user's watchlist after a successful login redirect.

Testing plan
============

1. Make sure you're on the mobile web view (Minerva).
2. Visit an article as a logged out/anon user.
3. Try to watch the article as an anon user.
4. MobileFrontend should tell you to login with various params in
   the URL
5. After you successfully login, the hook should fire and you should
   have the article automatically added to your watchlist.

Impact
======

End-users should not be impacted by this change. In fact, it should
be the same feel as before when we used UserLoginComplete hook.

Bug: T350138
Depends-On: Ie6e3fa24f4881855bd41c0361b8320c7f34918de
Change-Id: Ic874bca2a85c9da6679df094bbd0c5c2572a5f4e