Drop X-Large font size control
We are switching to three font-sizes to be consistent with desktop.
This adds code to migrate existing x-large preferences to large.
One side effect of this if you view Special:MobileOptions as your
first page view, no option will be selected but this is an edge case
that is rectified on page load so I reckon this is okay.

Bug: T349303
Change-Id: If5337c7346e0d7d482fa393e0ba2a7e1b5097823