First pass at replacing hardcoded colors with LESS variables
Replace hardcoded hex color values with their corresponding codex design
tokens.  The methodology for this was a regex search for `#[0-9a-f]+` in
the resources directory, excluding `resources/dist`

Not present in this patch is instances of `#72777d`, which we're still
figuring out what to do with, as well as a number of instances of `#000`
where `@color-emphasized` did not make sense.  Also excluded are fills
inside of SVGs, as I wasn't sure whether LESS variables would work
inside of them.  Lastly, two files need special attention - `mobilediff`
is using a number of custom colors not represented by any codex tokens,
and `ReferencesDrawer` is using a number of colors represented by
tokens, but "incorrectly" (border color for text, etc).  These will all
be addressed in a separate patch or patches

Note: I used my best judgement on a few of these, so I welcome a
thorough review if they feel questionable

Bug: T356427
Change-Id: Ifa8ace2c0b4aba81a74fd7b4c496825291e49173