Fixes search overlay icon
The new Codex mixin is not compatible with how we used to apply the icon on
input elements. This is corrected to use a pseudo element (at least for now)
that is overlayed on top of the input.

Note: after this change the Codex icon does grow on focus. A follow up will
be made in Minerva to keep these consistent.

This also fixes an issue with the input being unusable on night mode by
providing a color rather than relying on the browser default color.

Bug: T356427
Change-Id: I38c3c44084737062aab28165b5702779f97fd36c