Add class for test utils
Test classes should not be referenced from other test classes, because
they're not guaranteed to be loaded. Instead, code shared between test
classes should be located in non-test classes that are added to the
autoloader.

This allows MobileFrontendHooksTest to be run on its own, and also to
use paratest (T50217).

Change-Id: Ia2a133c2f9b87caa89a809b3c6cddab1cdc064d5