Improve the extension description
The description "Mobile Frontend" is too short and identical
to the extension name. Description are usually longer and
as their name implies, more descriptive.
At Wikimedia Hackathon 2024 I noticed that this confuses
translators, so I wanted to improve it.

I copied this description from the extension page
on mediawiki.org. It should be better for users
and translators.

Change-Id: I501b10dfe19fb4c7ecb23ca70c8b66b2ff1a2313