"mobile-frontend-copyright" is no longer a raw HTML message
Change I3f7b6f76d309e15a1976538206b17094b836addb in WikimediaMessages
removes the code that treated it as a raw HTML message. The remaining
implementation treats it as a wikitext message. The message text
doesn't contain any HTML or wikitext formatting anyway.

This message is not used in MobileFrontend, only in WikimediaMessages.
Maybe it should be moved there.

Depends-On: I3f7b6f76d309e15a1976538206b17094b836addb
Bug: T376293
Change-Id: I6df61c8ed6d18498683e2803e264678b4895970c