Make getLanguages() more robust
This function used to run on the assumption that there will
be a delimiter (–, en dash) between the page title and the
language name. However, since the title is translatable as a
MediaWiki message that can vary by language (as evidenced by [1]),
or even be customized on-wiki. Some languages use other delimiters,
some add extra characters, and the Japanese translation has the
language name *before* the title.

This change makes use of the new data attributes added in
Ib4bd0b8df2b79f07ceef16ff38d09ce02bed2367 which are not subject
to translation.

[1] https://translatewiki.net/wiki/Special:Translations/MediaWiki:Interlanguage-link-title

Bug: T349000
Change-Id: I30e5273d009d450ba607ef9ecddf51e59fec29eb