Don't overwrite DiscussionTools <span> tags during section collapsing xform
The contents of a heading tag when DiscussionTools is active are like:

```
 <span data-mw-comment-start="" id="h-Discussion_topic-20241126170200"></span>
 Discussion topic
 <span data-mw-comment-end="h-Discussion_topic-20241126170200"></span>
```

Be careful not to remove those <span> markers when adding the section
collapsing icons.

Change the nesting of the markup to better match the legacy parser.
This allows DiscussionTools to find its tags, and avoids disruption
to gadgets which use the ID to locate the heading.

Previous attempt: Ia9dfb88748defa6ec2610a82046fc7fd0e76cdfe.

Bug: T376048
Bug: T310544
Bug: T358981
Change-Id: I4467df9f09e04eda2a257a4886ab7ebd60c5e968