Emit Vary:X-Subdomain on index.php and api.php responses
When some part of the backend process may call shouldDisplayMobileView,
we must tell any middleware server, that a response for the current URL
can "vary", via the Vary header.

The primary call we care about is in,
MobileFrontendHooks::onRequestContextCreateSkin, where we set the skin
during index.php requests. But there are plenty other hooks and places
where MF, or an extension, may do the same.

Without this, an HTTP cache proxy will mix up responses for mobile
and desktop, such as Nginx, Varnish, or when using services like
Fastly or Cloudflare. The fix for this is to add custom configuration
to your cache proxy to vary on this fact manually, or, as WMF does,
by having separate m-dot hostnames, which mask this issue by changing
the full URL (i.e. the cache key) in a different way. This assumes
there will never be an m-dot request without X-Subdomain, or a
canonical request with X-Subdomain (which would still break the cache).

While at it:
* Remove hardcoded vary for X-Subdomain for onBeforePageRedirect.
  This is obsolete now. There were a dozen other hooks that also vary
  on shouldDisplayMobileView, but didn't add Vary.

Bug: T390929
Change-Id: I733d2b4f2ab7847e1090cfa3d280a81483b0ff25